pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.165.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
    "wordcloud>=1.9.4",
]
